Control Room Adjustments
Aligned the entrance, changed the door size b/c it was different on the z-scale for some reason. Elongated the hallway and moved 2 of the non-interactable doors further down the hallway. Moved the back wall that is supposed to be the holo screen a bit further back but im not sure if player can fit in the area behind