Comms and MedBay update
Added more lights to both levels. Added extra detail to the lower half of the medical bay. Contains individual rooms. Back 'wall' needs to be replaced with a transparant glass wall for narrative.
Block floating in the air next to one of the escape pods can be replaced with a broken escape pod.
Spikes blocking off the 2nd floor of the medical bay added.